// saudeConecta/src/app/util/variados/interfaces/consultaStatus/consultaStatus.ts
export interface ConsultaStatus {

  ConSttCodigoConsulata?: number;
  ConSttMedico?: number;
  ConSttPaciente?: number;
  ConSttDia_semana?: string;
  ConSttHorario?: string;
  ConSttData?: string;
  ConSttObservacao?: string;
  ConSttDataCriacao?: string;
  ConSttFormaPagamento?: number;
  ConSttStatus?: number;
  ConSttAdm?: number;




  conSttCodigoConsulata?: number;
  conSttMedico?: number;
  conSttPaciente?: number;
  conSttDia_semana?: string;
  conSttHorario?: string;
  conSttData?: string;
  conSttObservacao?: string;
  conSttDataCriacao?: string;
  conSttFormaPagamento?: number;
  conSttStatus?: number;
  conSttAdm?: number;
}